nyc-taxi-py/requirements.txt
azure-cosmos>=3.1.0
azure-functions
azure-functions-worker
azure-storage-file>=1.4.0
numpy>=1.16.3
python-dateutil>=2.8.0
